[libs] E: Adding nanvix-sandbox-cache Crate
