+++ src/libs/nanvix-sandbox-cache/Cargo.toml
@@ -1,0 +1,25 @@
+# Copyright(c) The Maintainers of Nanvix.
+# Licensed under the MIT License.
+
+[package]
+name = "nanvix-sandbox-cache"
+version.workspace = true
+license-file.workspace = true
+edition.workspace = true
+authors.workspace = true
+description = "Sandbox cache management for Nanvix"
+homepage.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+config = { workspace = true }
+hwloc = { workspace = true }
+nanvix-sandbox = { workspace = true }
+rand = { workspace = true }
+syscomm = { workspace = true }
+syslog = { workspace = true, features = ["std"] }
+tokio = { workspace = true, features = ["full"] }
+user-vm-api = { workspace = true }
+
+[features]
+default = []
