+++ src/libs/nanvix-sandbox-cache/src/lib.rs
@@ -1,9 +1,9 @@
 // Copyright(c) The Maintainers of Nanvix.
 // Licensed under the MIT License.
 
-//! Sandbox cache management.
+//! Sandbox cache management for Nanvix.
 //!
-//! This module provides caching functionality for sandboxed execution environments. It maintains
+//! This library provides caching functionality for sandboxed execution environments. It maintains
 //! a registry of active Linux Daemon and User VM instances, manages their lifecycle, and handles
 //! the control-plane socket connections for communication with these instances.
 
@@ -20,12 +20,20 @@ pub mod config;
 mod tag;
 
 //==================================================================================================
+// Exports
+//==================================================================================================
+
+pub use config::SandboxCacheConfig;
+pub use tag::SandboxTag;
+
+//==================================================================================================
 // Imports
 //==================================================================================================
 
-use crate::cache::{
-    config::SandboxCacheConfig,
-    tag::SandboxTag,
+use crate::config::{
+    control_plane_sockaddr_builder,
+    gateway_sockaddr_builder,
+    user_vm_sockaddr_builder,
 };
 use ::anyhow::Result;
 use ::nanvix_sandbox::{
@@ -82,6 +90,19 @@ pub struct SandboxCache {
     control_plane_socket: Option<Arc<Mutex<(SocketListener, String, SocketType)>>>,
     /// TCP port allocator for gateway ports in L2 deployment mode.
     tcp_port_allocator: TcpPortAllocator,
+    /// A builder function for control plane socket addresses.
+    control_plane_sockaddr_builder:
+        fn(&str, &str, bool) -> ::std::result::Result<String, ::anyhow::Error>,
+    /// A builder function for user VM socket addresses.
+    user_vm_sockaddr_builder:
+        fn(&str, &str, bool) -> ::std::result::Result<String, ::anyhow::Error>,
+    /// A builder function for gateway socket addresses.
+    gateway_sockaddr_builder: fn(
+        &str,
+        &str,
+        UserVmIdentifier,
+        &Option<TcpPort>,
+    ) -> ::std::result::Result<String, ::anyhow::Error>,
 }
 
 impl SandboxCache {
@@ -109,6 +130,9 @@ impl SandboxCache {
                 ::config::linuxd::GATEWAY_PORT_RANGE_BEGIN,
                 ::config::linuxd::GATEWAY_PORT_RANGE_END,
             ),
+            control_plane_sockaddr_builder,
+            user_vm_sockaddr_builder,
+            gateway_sockaddr_builder,
         }))
     }
 
@@ -181,17 +205,17 @@ impl SandboxCache {
                 };
 
                 // Work-out socket addresses before allocating any resources.
-                let control_plane_sockaddr: String = crate::config::control_plane_sockaddr_builder(
+                let control_plane_sockaddr: String = (self.control_plane_sockaddr_builder)(
                     self.config.tmp_directory(),
                     tag.tenant_id(),
                     self.config.l2(),
                 )?;
-                let user_vm_sockaddr: String = crate::config::user_vm_sockaddr_builder(
+                let user_vm_sockaddr: String = (self.user_vm_sockaddr_builder)(
                     self.config.tmp_directory(),
                     tag.tenant_id(),
                     self.config.l2(),
                 )?;
-                let gateway_sockaddr: String = crate::config::gateway_sockaddr_builder(
+                let gateway_sockaddr: String = (self.gateway_sockaddr_builder)(
                     self.config.tmp_directory(),
                     tag.tenant_id(),
                     tag.sandbox_id(),
