+++ src/utils/nanvixd/Cargo.toml
@@ -25,6 +25,7 @@ linuxd = { workspace = true }
 num_enum = { workspace = true }
 rand = { workspace = true }
 nanvix-sandbox = { workspace = true }
+nanvix-sandbox-cache = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true, features = ["std"] }
 sys = { workspace = true }
